pyproject.toml
[project]
name = "prom-test"
version = "0.1.0"
description = "Testing infrastructure for the DPDK prom feature"
requires-python = ">=3.10"
dependencies = [
    "pytest>=8.0",
    "pyyaml>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"
